♻️ refactor(web): extract layout init, modals, and shortcuts
