+++ apps/web/package.json
@@ -1,7 +1,7 @@
 {
   "name": "web",
   "private": true,
-  "version": "0.16.24",
+  "version": "0.16.27",
   "type": "module",
   "scripts": {
     "dev": "vite dev",
