+++ apps/web/src/service-worker.ts
@@ -5,7 +5,7 @@
 
 import { build, files, version } from "$service-worker";
 
-const CACHE_VERSION = "213";
+const CACHE_VERSION = "216";
 const CACHE = `cache-${version}-${CACHE_VERSION}`;
 
 const ASSETS = [
